Move Rebuilt field constants to main project folder
